Refactor back to original project structure
